cleanup token contracts main
